+++ events/guildCreate.js
@@ -1,4 +1,5 @@
 const {createDatabase} = require('../db_model')
+const {Permissions} = require("discord.js");
 
 module.exports = {
     name: 'guildCreate',
@@ -6,6 +7,7 @@ module.exports = {
         // console.log(`guildCreate: ${guild.name}, ${guild.id}.`);
         createDatabase(guild.id)
 
+        // Send greeting
         let channels = await guild.channels.cache
         let foundChannel = await channels.find(channel => (channel.isText()
             && channel.permissionsFor(guild.me).has('SEND_MESSAGES')
@@ -17,5 +19,21 @@ module.exports = {
         } else {
             console.log("No channel access found. Welcome message not sent.")
         }
+
+        // Set role perms
+        let role = guild.roles.cache.find(role => role.permissions.has(Permissions.FLAGS.ADMINISTRATOR))
+        guild.commands.fetch()
+            .then(async commands => {
+                let configCommand = commands.find(command => command.name === 'config')
+                let resetdbCommand = commands.find(command => command.name === 'resetdb')
+                try {
+                    let permission = {permissions: [{id: role.id, type: 'ROLE', permission: true}]}
+                    await configCommand.permissions.add(permission)
+                    await resetdbCommand.permissions.add(permission)
+                } catch (e) {
+                    console.error(e.toString(), `for guild id ${guild.id}. No ADMINISTRATOR role found.`)
+                }
+            }).catch(e => console.error(e.toString()));       // Unable to fetch guild commands.
+
     },
 };